removed depricated doubel

+++ faker/src/test/java/com/github/faker/matchers/IsANumber.java
@@ -4,12 +4,14 @@ import org.hamcrest.Description;
 import org.hamcrest.Matcher;
 import org.hamcrest.TypeSafeMatcher;
 
+import java.math.BigDecimal;
+
 public class IsANumber extends TypeSafeMatcher<String> {
 
     @Override
     protected boolean matchesSafely(String item) {
         try {
-            new Double(item);
+            new BigDecimal(item);
         } catch (NumberFormatException nfe) {
             return false;
         }
